Refactor registration section: - Updated FormField styles for improved UI. - Removed unused PaymentWompi.css and index.tsx files. - Introduced WhatsAppWidget and WA-button components for better user interaction. - Created Wompi-button component for payment processing. - Added thank you page and integrated WhatsApp widget. - Updated main page to include ReserveInfo instead of RegistrationForm. - Changed CTA section ID for better clarity. - Added thank you image asset.
