--- app/components/sections/registration/PaymentWompi.css
@@ -1,4 +1,0 @@
-div.waybox-modal,
-div.waybox-display-block {
-  position: fixed;
-}
