+++ app/page.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Hero from '@/app/components/sections/Hero';
 import Header from '@/app/components/layout/Header';
 import Footer from '@/app/components/layout/Footer';
-import RegistrationForm from '@/app/components/sections/registration';
+import ReserveInfo from '@/app/components/sections/ReserveInfo';
 import TargetAudience from '@/app/components/sections/TargetAudience';
 import ExperienceHighlights from '@/app/components/sections/ExperienceHighlights';
 import TransformationEnvironment from '@/app/components/sections/TransformationEnvironment';
@@ -11,11 +11,11 @@ import PriceCard from '@/app/components/sections/Price';
 import Testimonials from '@/app/components/sections/Testimonials';
 import FAQ from '@/app/components/sections/FAQ';
 import YourGuide from '@/app/components/sections/YourGuide';
-import WhatsAppWidget from '@/app/components/ui/WhatsAppWidget';
+import WhatsAppWidget from '@/app/components/ui/WA-Widget';
 
 export default function Home() {
   return (
-    <main className="min-h-screen bg-white">
+    <main className="min-h-screen bg-dark-bg">
       <Header />
       <Hero />
       <TargetAudience />
@@ -24,13 +24,10 @@ export default function Home() {
       <Testimonials />
       <Benefits />
       <PriceCard />
-      <RegistrationForm />
+      <ReserveInfo />
       <YourGuide />
       <FAQ />
       <Footer />
-
-      {/* Sticky WhatsApp Widget */}
-      <WhatsAppWidget />
     </main>
   );
 }
