+++ app/variables/index.ts
@@ -37,7 +37,7 @@ export const experience = {
 
 export const eventDate = new Date("2025-06-14T00:00:00");
 
-export const CTA_SECTION_ID = "precio";
+export const CTA_SECTION_ID = "registration-form";
 
 export const whatsappContact = {
   number: "+573127811615", // Formatted for WhatsApp (removed spaces)
